+++ src/pages/mypage/MyPageMain.tsx
@@ -2,34 +2,15 @@ import styled from 'styled-components';
 import InvitationStats from '../../components/mypage/InvitationStats';
 import { removeCookie } from '../../utils/cookies';
 import { useNavigate } from 'react-router-dom';
-import { useEffect, useState } from 'react';
-import { useResetRecoilState } from 'recoil';
-import { UserInfo } from '../../atom/UserInfo';
-import { useErrorBoundary } from 'react-error-boundary';
-import { myPageInfoAPI } from '../../api';
+import { useGetMyPageInfo } from '../../api/useGetMyPageInfo';
 
 const MyPageMain = () => {
-  const resetUserInfo = useResetRecoilState(UserInfo);
-  const { showBoundary } = useErrorBoundary();
+  const { user } = useGetMyPageInfo();
   const navigate = useNavigate();
   const handleLogout = () => {
     removeCookie('accessToken');
     navigate('/', { replace: true });
   };
-
-  const [user, setUser] = useState({
-    nickname: '',
-    receivedInvitationCount: 0,
-    sentInvitationCount: 0,
-  });
-
-  useEffect(() => {
-    myPageInfoAPI(resetUserInfo, showBoundary).then((res) => {
-      if (res.isSuccess && res.result !== null) {
-        setUser(res.result);
-      }
-    });
-  }, []);
 
   return (
     <UserBox>
